add test for SPI
