+++ plugins/autodoc/autodoc-processor/src/test/java/org/eclipse/dataspaceconnector/plugins/autodoc/core/processor/testextensions/SecondExtension.java
@@ -14,7 +14,6 @@
 
 package org.eclipse.dataspaceconnector.plugins.autodoc.core.processor.testextensions;
 
-import org.eclipse.dataspaceconnector.plugins.autodoc.core.processor.SomeService;
 import org.eclipse.dataspaceconnector.runtime.metamodel.annotation.Inject;
 
 public class SecondExtension {
