+++ plugins/autodoc/autodoc-processor/src/test/java/org/eclipse/dataspaceconnector/plugins/autodoc/core/processor/testextensions/SomeOtherService.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.plugins.autodoc.core.processor;
+package org.eclipse.dataspaceconnector.plugins.autodoc.core.processor.testextensions;
 
 public interface SomeOtherService {
 }
